requirements.txt
python-telegram-bot==21.6
yt-dlp>=2024.11.18
python-dotenv==1.0.1
